requirements.txt
discord.py
discord-ext-voice-recv @ git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git
python-dotenv
aiohttp
PyNaCl
python-telegram-bot
edge-tts
SpeechRecognition
pydub
